requirements.txt
insightface
opencv-python
videostream
faiss-cpu
imutils
onnxruntime
redis
